pyproject.toml
[project]
name = "sumtablets"
version = "0.1.0"
description = "Sumerian cuneiform ↔ transliteration parallel corpus pipeline"
authors = [{ name = "Cole Simmons", email = "colesimmons223@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.8.2",
    "pandas>=2.2.2",
    "requests>=2.32.3",
    "scikit-learn>=1.5.0",
    "tqdm>=4.66.4",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sumtablets"]
